requirements.txt
discord.py[voice]>=2.1.1, <3.0.0
yt-dlp>=2023.02.17
setuptools